playground/pyproject.toml
[project]
name = "playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pycrypto>=2.6.1",
]
